select all action of the menu bar is connected to the interactor select elements
