+++ src/negui_main_widget.h
@@ -1,0 +1,49 @@
+#ifndef __NEGUI_MAINWIDGET_H
+#define __NEGUI_MAINWIDGET_H
+
+main dddd#include <QObject>
+#include <QWidget>
+#include <QFrame>
+
+class MyNetworkEditorWidget : public QFrame {
+    Q_OBJECT
+    
+public:
+    
+    explicit MyNetworkEditorWidget(QWidget *parent = nullptr);
+    ~MyNetworkEditorWidget();
+
+    QObject* interactor();
+    QWidget* title();
+    QWidget* toolBar();
+    QWidget* modeMenu();
+    QWidget* view();
+    void readSettings();
+    void writeSettings();
+
+signals:
+
+    void askForSetNewNetworkCanvas();
+    void askForSelectAll();
+
+private slots:
+
+    void displayFeatureMenu(QWidget* featureMenu);
+    void removeFeatureMenu();
+
+protected:
+    
+    void setWidgets();
+    void setInteractions();
+    void setReadyToLaunch();
+    void closeEvent(QCloseEvent *event) override;
+
+    QObject* _interactor;
+    QWidget* _title;
+    QWidget* _toolBar;
+    QWidget* _view;
+    QWidget* _modeMenu;
+    QWidget* _featureMenu;
+};
+
+#endif
